+++ include/matrix_tools.hpp
@@ -13,7 +13,7 @@ public:
     int nb_cols;
 
     kMatrix()
-    : buffer(std::vector<double>(3)),
+    : buffer(std::vector<float>(3)),
     nb_rows(3),
     nb_cols(1)
     {
@@ -39,13 +39,13 @@ public:
 //Eigen::Matrix4d invert_matrix4(const Eigen::Matrix4d& mat);
 //Eigen::Matrix3d invert_matrix3(const Eigen::Matrix3d &mat);
 
-std::vector<double> invert_matrix(const std::vector<double> &mat, size_t len);
+std::vector<float> invert_matrix(const std::vector<float> &mat, size_t len);
 
-double compute_det(const std::vector<double> &mat, size_t len);
+float compute_det(const std::vector<float> &mat, size_t len);
 
-//Eigen::MatrixXd vec_to_eigen(std::vector<double> vec, size_t n_row, size_t n_col);
+//Eigen::MatrixXd vec_to_eigen(std::vector<float> vec, size_t n_row, size_t n_col);
 
-std::vector<double> matmul(const std::vector<double> &lhs, const std::vector<double> &rhs, size_t n, size_t p,
+std::vector<float> matmul(const std::vector<float> &lhs, const std::vector<float> &rhs, size_t n, size_t p,
                            size_t m);
 //
 //Eigen::MatrixXd operator%(const Eigen::MatrixXd &lhs, const Eigen::MatrixXd &rhs);
@@ -65,12 +65,12 @@ kMatrix<T> operator+(const kMatrix<T> &lhs, const kMatrix<T> &rhs)
     {
         throw std::runtime_error("Can't add matrices with different dimensions");
     }
-    auto out_buffer = std::vector<double>(lhs.nb_cols * lhs.nb_rows);
+    auto out_buffer = std::vector<float>(lhs.nb_cols * lhs.nb_rows);
     for (int i = 0; i < lhs.nb_cols * lhs.nb_rows; i++)
     {
         out_buffer[i] = lhs.buffer[i] + rhs.buffer[i];
     }
-    return kMatrix<double>(out_buffer, lhs.nb_rows, lhs.nb_cols);
+    return kMatrix<float>(out_buffer, lhs.nb_rows, lhs.nb_cols);
 }
 
 template <typename T>
@@ -80,19 +80,19 @@ kMatrix<T> operator-(const kMatrix<T> &lhs, const kMatrix<T> &rhs)
     {
         throw std::runtime_error("Can't add matrices with different dimensions");
     }
-    auto out_buffer = std::vector<double>(lhs.nb_cols * lhs.nb_rows);
+    auto out_buffer = std::vector<float>(lhs.nb_cols * lhs.nb_rows);
     for (int i = 0; i < lhs.nb_cols * lhs.nb_rows; i++)
     {
         out_buffer[i] = lhs.buffer[i] - rhs.buffer[i];
     }
-    return kMatrix<double>(out_buffer, lhs.nb_rows, lhs.nb_cols);
+    return kMatrix<float>(out_buffer, lhs.nb_rows, lhs.nb_cols);
 }
 
 template <typename T>
 kMatrix<T> invert_matrix3(const kMatrix<T> &mat)
 {
     auto inv_buf = invert_matrix(mat.buffer, 3);
-    return kMatrix<double>(inv_buf, 3, 3);
+    return kMatrix<float>(inv_buf, 3, 3);
 }
 
 template <typename T>
@@ -110,4 +110,4 @@ kMatrix<T> matmul(const kMatrix<T> &lhs, const kMatrix<T> &rhs)
 }
 
 
-std::vector<double> get_adjugate_matrix(const std::vector<double> &mat, size_t len);
+std::vector<float> get_adjugate_matrix(const std::vector<float> &mat, size_t len);
